requirement.txt
gradio==3.36.1
datasets==2.14.5
chromadb==0.3.23
pillow
numpy
openai=0.28.0
matplotlib
python-dotenv
huggingface_hub
pyttsx3